move app wrapper to own file
